+++ packages/graphql/tests/utils/builders/relation-field-builder.ts
@@ -18,6 +18,7 @@
  */
 
 import { RelationshipQueryDirectionOption } from "../../../src/constants";
+import { defaultNestedOperations } from "../../../src/graphql/directives/relationship";
 import type { RelationField } from "../../../src/types";
 import { Builder } from "./builder";
 
@@ -37,6 +38,7 @@ export class RelationFieldBuilder extends Builder<RelationField, RelationField> 
             arguments: [],
             inherited: false,
             queryDirection: RelationshipQueryDirectionOption.DEFAULT_DIRECTED,
+            nestedOperations: defaultNestedOperations,
             ...newOptions,
         });
     }